requirements.txt
django-bootstrap-form
torch
torchvision
Pillow
psycopg2
sklearn
